add app-containers/podman-compose ebuilds
for the latest two versions available. 1.0.6 seems to be very new
compared to the next recent 1.0.3 which was the latest version for a
long time, so it felt like a good idea to have both available just
in case

also i still feel like i have very little idea what i'm doing with
ebuild writing, and even less so with python stuff. this is all feeling
like a lot of "magic" to me. hopefully this isn't written in a
horribly broken or bad way ...